requirements/dev.txt
-r common.txt
ipython
django-nose<1.5
django-extensions
ipdb
Faker
